Many changes.
Beginning of log scale for Plots.  Changed parameters will wake and
animate Job. Add comments everywhere.  Remove unused Ordering2App.

At some point I want to revisit GUIValues, because its kind of ugly how the callbacks are handled.  Every time a G'Val is changed, it calls the change callback, which "invokes later" from the swing thread.  this is kind of overkill, sometimes you're already in the swing thread.  the point is that different classes of calbacks are lumped together.  for example, when the parameters are locked, each one triggers all callbacks, invoking job.wake() many times, when the correct number should be zero. So i guess the thing to do is have multiple types of change callbacks.. ?